requirements - Copy.txt
streamlit
google-generativeai
Pillow
PyPDF2